% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_annotations.R
\name{append_annotations}
\alias{append_annotations}
\title{Append annotation data frames with automatic column alignment}
\usage{
append_annotations(data.to.append, existing.annotations)
}
\arguments{
\item{data.to.append}{A data frame of new annotations to be appended.}

\item{existing.annotations}{A data frame of existing annotations.}
}
\value{
A combined data frame containing both the new and existing
annotations with aligned columns.
}
\description{
This function appends a new annotation data frame to an existing
annotation table. If the new data contains columns that are not
present in the existing annotations, those columns are added to
the existing annotations and filled with \code{NA}.
}
\examples{
\dontrun{
new <- data.frame(feature.ID = 1:2, compound.name = c("A", "B"))
old <- data.frame(feature.ID = 3, compound.name = "C", confidence.score = 0.95)
combined <- append_annotations(new, old)
}

}
