% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_propagated_annotations.R
\name{append_propagated_annotations}
\alias{append_propagated_annotations}
\title{Append Propagated Annotations to a Data Frame}
\usage{
append_propagated_annotations(
  full.annotation.data,
  propagated_df,
  mod_db = modification_db,
  ppm_tol = 10,
  abs_tol = 0.01
)
}
\description{
Merges propagated annotations into the main annotation data frame,
calculates mass differences to identify specific modifications, and updates compound names.
}
