% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gnps_info.R
\name{extract_gnps_info}
\alias{extract_gnps_info}
\title{Extract GNPS Information from Metadata}
\usage{
extract_gnps_info(excel.metadata.file, dataset.id, output.dir = "HGM")
}
\arguments{
\item{excel.metadata.file}{A character string. The path to the Excel metadata file.}

\item{dataset.id}{A character string. The ID used to filter the dataset in the metadata.}

\item{output.dir}{A character string. The directory where CSV files will be saved. Defaults to "HGM".}
}
\value{
A list containing two elements: \code{gnps.task.id} (a character string) and \code{rt.tol} (a numeric value).
}
\description{
This function reads an Excel metadata file, converts its sheets to
CSV files, and extracts a GNPS task ID and retention time tolerance
from the resulting data.
}
