% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cid_with_fallbacks.R
\name{get_cid_only_with_fallbacks}
\alias{get_cid_only_with_fallbacks}
\title{Get CID from PubChem with Fallbacks and LipidMaps Lookup}
\usage{
get_cid_only_with_fallbacks(
  name,
  smiles = NA,
  cid_cache_df,
  lipids.file,
  offline = FALSE
)
}
\arguments{
\item{name}{A character string representing the compound name.}

\item{smiles}{A character string representing the compound SMILES (optional).}

\item{cid_cache_df}{The data frame to be used as a cache, with columns LookupName and CID.}

\item{lipids.file}{A data frame to use to lookup lipid names for PubChem CID lookups.}

\item{offline}{Logical; if TRUE, strictly avoids live PubChem API lookups.}
}
\value{
A named list with keys: \code{CID} and the updated \code{cid_cache_df}.
}
\description{
Attempts to resolve a PubChem CID using compound name and SMILES, checking LipidMaps first,
and updates a provided cache which only stores the CID.
}
