% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPS.R
\name{MAPS}
\alias{MAPS}
\title{Metabolomics Annotation Processing and Standardisation (MAPS)}
\usage{
MAPS(
  dataset.id,
  folder,
  cache.location,
  lipids.location,
  gnps.task.id,
  gnps.prob = 0.7,
  canopus.prob = 0.7,
  csi.prob = 0.64,
  ms2query.prob = 0.7,
  ppm.tol = 5,
  rt.tol = 0.1,
  cid_database_path = NULL,
  standardisation,
  lv1.subclasses,
  lv2.mzmine,
  msnovelist,
  modification_db,
  updateProgress = NULL,
  enable_local_db = TRUE,
  enable_api = TRUE
)
}
\arguments{
\item{dataset.id}{Character string identifying the dataset (for file naming).}

\item{folder}{Character string. \strong{Required}: Path to the specified Processed Data Folder containing results from MZmine, GNPS, SIRIUS, etc.}

\item{gnps.task.id}{Character string. \strong{Required}: External GNPS ID for the task (can be extracted from metadata).}

\item{gnps.prob}{Numeric. Minimum probability/score for high-confidence GNPS annotations (default: 0.7).}

\item{canopus.prob}{Numeric. Minimum probability for CANOPUS (NPC Pathway) annotations (default: 0.7).}

\item{csi.prob}{Numeric. Minimum probability/score for CSI:FingerID annotations (default: 0.64).}

\item{ms2query.prob}{Numeric. Minimum probability/score for MS2Query annotations (default: 0.7).}

\item{rt.tol}{Numeric. Retention time tolerance, e.g., for C18 column (default: 0.1).}

\item{cid_database_path}{Character string. Path to the indexed PubChem SQLite database (default: "Y:/MA_BPA_Microbiome/Databases/PubChem/PubChem_Indexed.sqlite").}

\item{updateProgress}{Function. A callback function to update a Shiny progress bar (optional).}
}
\value{
An invisible data frame containing the final, processed, and propagated annotations.
}
\description{
This is the main function for processing and integrating
metabolomics annotation data from various sources (GNPS, SIRIUS, MS2Query).
}
