% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cytoscape_file.R
\name{prepare_cytoscape_file}
\alias{prepare_cytoscape_file}
\title{Prepare and Append Cytoscape File}
\usage{
prepare_cytoscape_file(
  folder,
  propagated.annotation.data.with.samples,
  read_checked_csv,
  write_large_csv
)
}
\arguments{
\item{folder}{The base folder path for file input/output.}

\item{propagated.annotation.data.with.samples}{The data frame containing propagated annotations and samples.}

\item{read_checked_csv}{A function to read checked CSV files.}

\item{write_large_csv}{A function to write large CSV files.}
}
\value{
The processed Cytoscape data frame.
}
\description{
Reads the GNPS Cytoscape file, appends the best annotations, and writes the final file to disk.
}
