% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_append_sample_data.R
\name{process_and_append_sample_data}
\alias{process_and_append_sample_data}
\title{Process and Append Sample Data}
\usage{
process_and_append_sample_data(sample.data, propagated.annotation.data)
}
\arguments{
\item{sample.data}{A data frame containing sample information, including peak areas.}

\item{propagated.annotation.data}{A data frame of existing annotations.}
}
\description{
Cleans and processes sample data to create a presence/absence list,
then appends it to a main annotation data frame.
}
