% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_canopus_data.R
\name{process_canopus_data}
\alias{process_canopus_data}
\title{Process CANOPUS Data}
\usage{
process_canopus_data(canopus.data_path)
}
\arguments{
\item{canopus.data_path}{Path to the CANOPUS annotations TSV file.}
}
\value{
A data frame containing the processed CANOPUS annotations.
}
\description{
Loads, cleans, and processes CANOPUS annotations from a TSV file.
}
