% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gnps_data.R
\name{process_gnps_data}
\alias{process_gnps_data}
\title{Process GNPS annotation and cluster data}
\usage{
process_gnps_data(gnps.task.id, gnps.prob, mzmine.annotations.final)
}
\arguments{
\item{gnps.task.id}{GNPS task ID (character)}

\item{gnps.prob}{Minimum confidence score (numeric)}

\item{mzmine.annotations.final}{Data frame of MZmine annotations}
}
\value{
A list with five data frames including levels 2 and 3 annotations and networking data.
}
\description{
Loads, tidies, and standardizes GNPS2 annotation data, and splits into
Level 2 high-confidence, Level 2 low-confidence, and Level 3 annotation tables.
}
