% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mzmine_DATA.R
\name{process_mzmine_data}
\alias{process_mzmine_data}
\title{Process MZMINE Data}
\usage{
process_mzmine_data(
  mzmine.annotations,
  gnps.prob,
  cid.cache.path = cache.location,
  lipids.file.path = lipids.location,
  cid.database.path = NULL,
  standardisation,
  enable_local_db = enable_local_db,
  enable_api = enable_api
)
}
\arguments{
\item{mzmine.annotations}{Path to the MZMINE annotations CSV file.}

\item{gnps.prob}{A numeric value for the GNPS confidence probability threshold.}

\item{cid.cache.path}{Path to the CID cache CSV file. Defaults to \verb{~/cid_cache.csv}.}

\item{lipids.file.path}{Path to the lipids TSV file. Defaults to a specific network path.}

\item{cid.database.path}{Path to the PubChem CID SQLite database. Defaults to a specific network path.}

\item{mzmine.data}{Path to the MZMINE data CSV file.}
}
\value{
A list containing four data frames: \code{processed.data}, \code{lipids.data}, \code{annotations.data}, and \code{initial.mzmine.data}.
}
\description{
A function to load, process, and enrich MZMINE annotation data.
}
