% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_zodiac_data.R
\name{process_zodiac_data}
\alias{process_zodiac_data}
\title{Process ZODIAC Data}
\usage{
process_zodiac_data(zodiac.data_path)
}
\arguments{
\item{zodiac.data_path}{Path to the ZODIAC annotations TSV file.}
}
\value{
A data frame containing the processed ZODIAC annotations.
}
\description{
Loads, cleans, and processes ZODIAC annotations from a TSV file.
}
