% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_with_cache.R
\name{standardise_with_cache}
\alias{standardise_with_cache}
\title{Standardise Compound Annotations with Caching}
\usage{
standardise_with_cache(
  df,
  name_col,
  smiles_col,
  cid_cache_df,
  lipids.file,
  db_path
)
}
\arguments{
\item{df}{A data frame containing compound annotations.}

\item{name_col}{Name of the column in \code{df} containing compound names (as a string).}

\item{smiles_col}{Name of the column in \code{df} containing SMILES strings (as a string).}

\item{cid_cache_df}{A data frame to use as a cache for PubChem CID lookups. This cache will be updated.}

\item{lipids.file}{A data frame to use to lookup lipid names for PubChem CID lookups.}

\item{db_path}{Path to the "SQLite database" file containing PubChem data.}
}
\value{
A list with two elements: \code{data} (the updated data frame) and \code{cache} (the updated CID cache).
}
\description{
This function acts as a wrapper for \code{standardise_annotation}, providing a
convenient interface to standardise compound names and SMILES strings
in a data frame. It resolves CIDs using a local cache and PubChem,
then updates the data with properties from a local SQLite database.
}
