% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_final_files.R
\name{write_final_files}
\alias{write_final_files}
\title{Write Final Processed Files}
\usage{
write_final_files(
  final.annotation.df,
  samples.df,
  folder,
  dataset.id,
  mzmine.data,
  cid_cache_df,
  write_large_csv
)
}
\arguments{
\item{final.annotation.df}{The main processed annotation data frame.}

\item{samples.df}{The processed samples data frame.}

\item{folder}{The base folder path for file output.}

\item{dataset.id}{The unique identifier for the dataset.}

\item{mzmine.data}{The initial MzMine data frame for verification.}

\item{cid_cache_df}{The CID cache data frame to be saved.}

\item{write_large_csv}{A function to write large CSV files.}
}
\value{
A message indicating success or an error with processing time.
}
\description{
This function takes the final processed data frames and writes them to disk in specified locations. It also performs a final data verification check.
}
