% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_annotations.R
\name{append_annotations}
\alias{append_annotations}
\title{Append Level 4 and 5 Annotations}
\usage{
append_annotations(
  data,
  mask_condition,
  compound_name,
  confidence_level,
  npc_pathway,
  npc_superclass,
  annotation_type
)
}
\arguments{
\item{data}{The data frame to be annotated.}

\item{mask_condition}{A logical vector (or an expression that evaluates to one) indicating which rows to update.}

\item{compound_name}{The new compound name to be assigned.}

\item{confidence_level}{The confidence level to be assigned.}

\item{npc_pathway}{The NPC pathway to be assigned.}

\item{npc_superclass}{The NPC superclass to be assigned.}

\item{annotation_type}{The annotation type to be assigned.}
}
\value{
The data frame with the new annotations appended.
}
\description{
Appends annotations based on a provided mask and updates relevant columns.
}
