% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ms2query_analogues.R
\name{append_ms2query_analogues}
\alias{append_ms2query_analogues}
\title{Format and Append MS2Query Analogue Annotations}
\usage{
append_ms2query_analogues(
  ms2query_data,
  existing_annotations,
  mod_db,
  abs_tol = 0.01
)
}
\arguments{
\item{ms2query_data}{A data frame containing MS2Query Level 3 annotations.}

\item{existing_annotations}{A data frame of existing annotations.}

\item{mod_db}{The modification dataframe.}

\item{abs_tol}{The absolute mass tolerance in Daltons (default 0.01).}
}
\value{
The updated MS2Query data frame, formatted and ready for merging.
}
\description{
Processes MS2Query Level 3 analogue data, maps the existing mz.diff against a modification database, unconditionally formats the compound names, and prepares the data for merging.
}
