% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_propagated_annotations.R
\name{append_propagated_annotations}
\alias{append_propagated_annotations}
\title{Append Propagated Annotations to a Data Frame}
\usage{
append_propagated_annotations(full.annotation.data, propagated_df)
}
\arguments{
\item{full.annotation.data}{The original full annotation data frame.}

\item{propagated_df}{The data frame of newly propagated annotations.}
}
\value{
A single data frame with the propagated annotations appended.
}
\description{
Merges propagated annotations into the main annotation data frame and updates relevant columns.
}
