% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_compound_names.R
\name{fix_compound_names}
\alias{fix_compound_names}
\title{Clean and Standardize Compound Names}
\usage{
fix_compound_names(df, column)
}
\arguments{
\item{df}{A data.frame or data.table with compound names.}

\item{column}{A character vector of column names to clean.}
}
\value{
A cleaned data frame.
}
\description{
Applies regex-based cleanup and casing corrections to specified column(s) of a data frame.
}
