% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_append_data.R
\name{merge_and_append_data}
\alias{merge_and_append_data}
\title{Merge and Append Data Frames}
\usage{
merge_and_append_data(new_data, existing_annotations)
}
\arguments{
\item{new_data}{A data frame containing new annotations to be appended.}

\item{existing_annotations}{A data frame of existing annotations.}
}
\value{
The combined data frame with new annotations appended.
}
\description{
Merges a new data frame of annotations with an existing one, handling column discrepancies.
}
