% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_append_msnovelist.R
\name{process_and_append_msnovelist}
\alias{process_and_append_msnovelist}
\title{Process and Append MSNovelist Annotations}
\usage{
process_and_append_msnovelist(
  msn.data,
  existing_annotations,
  msn.threshold,
  cid_cache_df,
  lipids.file,
  cid_database_path,
  compute_id_prob,
  deduplicate_data,
  standardise_annotation,
  standardisation
)
}
\arguments{
\item{msn.data}{Path to the MSNovelist tsv file.}

\item{existing_annotations}{A data frame of existing annotations.}

\item{msn.threshold}{A numeric threshold for MSNovelist ModelScore.}

\item{cid_cache_df}{A data frame used for caching CID lookups.}

\item{lipids.file}{Path to a lipids file for standardization.}

\item{cid_database_path}{Path to the PubChem CID database.}

\item{compute_id_prob}{A function to compute identification probability.}

\item{deduplicate_data}{A function to remove duplicate annotations.}

\item{standardise_annotation}{A function to standardize annotations.}
}
\value{
A list containing the updated annotations data frame and the updated CID cache.
}
\description{
Cleans, standardizes, and appends MSNovelist de novo annotations to an existing data frame.
}
