% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy_fixer.R
\name{redundancy_fixer}
\alias{redundancy_fixer}
\title{Identify Redundant Features Based on 3-Tier Hierarchy and Retention Time}
\usage{
redundancy_fixer(dataset, rt_column = "rt", rt_tolerance = 1)
}
\arguments{
\item{dataset}{A data frame containing the data to process.}

\item{rt_column}{Column name (as string) indicating retention time.}

\item{rt_tolerance}{Numeric value specifying the maximum allowable RT difference.}
}
\value{
The original dataset with a new logical column \code{redundant}, indicating flagged redundant rows.
}
\description{
Flags potentially redundant entries in a dataset based on:
\enumerate{
\item CID
\item Identical compound.name
\item A combination of compound.name AND smiles
It never checks 'smiles' exclusively.
}
}
