% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_annotation.R
\name{standardise_annotation}
\alias{standardise_annotation}
\title{Standardise Compound Annotations Using PubChem}
\usage{
standardise_annotation(
  data,
  name_col = "compound_name",
  smiles_col = "smiles",
  cid_cache_df = NULL,
  lipids.file = NULL,
  cid_database_path = NULL,
  standardisation = standardisation
)
}
\arguments{
\item{data}{A data frame containing compound annotations.}

\item{name_col}{Name of the column in \code{data} containing compound names (as a string).}

\item{smiles_col}{Name of the column in \code{data} containing SMILES strings (as a string).}

\item{cid_cache_df}{A data frame to use as a cache for PubChem CID lookups.}

\item{lipids.file}{A data frame to use to lookup lipid names for PubChem CID lookups.}

\item{cid_database_path}{Path to the "SQLite database" file.}
}
\value{
A list with two elements: \code{data} (the updated data frame) and \code{cache} (the updated CID cache).
}
\description{
For each row in a data frame, attempts to resolve compound CIDs using PubChem.
Then, uses the resolved CIDs to retrieve and update compound name (with PubChem Title),
SMILES, Formula, IUPAC, and Monoisotopic.Mass from a local SQLite database.
Rows containing "candidate" (case-insensitive) in the name column are excluded.
NOTE: The function is modified to proceed even if the local SQLite database is missing,
by skipping the property retrieval step (Pass 2).
}
